+++ src/kernel/timer.h
@@ -9,6 +9,8 @@ void timer_handle_irq(void);
 unsigned int timer_get_ticks(void);
 unsigned int timer_get_seconds(void);
 unsigned int timer_get_hz(void);
+unsigned int timer_get_realtime_seconds(void);
+void timer_set_realtime_seconds(unsigned int seconds);
 unsigned int timer_ms_to_ticks_round_up(unsigned int ms);
 
 #endif